+++ src/Device/deviceLister.ts
@@ -11,13 +11,13 @@ import nrfDeviceLib, {
 } from '@nordicsemiconductor/nrf-device-lib-js';
 
 import logger from '../logging';
-import type { AppDispatch, RootState } from '../store';
+import type { AppThunk, RootState } from '../store';
 import {
-    clearWaitForDevice,
     clearWaitForDeviceTimeout,
     setArrivedButWrongWhen,
     setDisconnectedTime,
     setLastArrivedDeviceId,
+    setOnCancelTimeout,
     setWaitForDeviceTimeout,
     WaitForDevice,
 } from './deviceAutoSelectSlice';
@@ -72,10 +72,11 @@ const shouldAutoReselect = (
 };
 
 const initAutoReconnectTimeout =
-    (onTimeout: () => void, waitForDevice?: WaitForDevice) =>
-    (dispatch: AppDispatch) => {
-        const timeout = waitForDevice?.timeout;
-        if (timeout == null) return;
+    (onTimeout: () => void, waitForDevice: WaitForDevice): AppThunk =>
+    dispatch => {
+        const timeout = waitForDevice.timeout;
+
+        dispatch(setOnCancelTimeout(onTimeout));
 
         dispatch(
             setWaitForDeviceTimeout(
@@ -87,8 +88,7 @@ const initAutoReconnectTimeout =
                                 timeout / 1000
                             } seconds.`
                         );
-                    dispatch(clearWaitForDeviceTimeout());
-                    onTimeout();
+                    dispatch(clearWaitForDevice());
                     logger.warn(
                         `Failed to detect device after reboot. Timed out after ${
                             timeout / 1000
@@ -136,6 +136,33 @@ export const hasValidDeviceTraits = (
         rule => requiredTraits[rule as keyof DeviceTraits] === false
     );
 
+const removeDeviceFromList =
+    (
+        removedDevice: Device,
+        onDeviceDeselected: () => void,
+        onDeviceDisconnected: (device: Device) => void
+    ): AppThunk =>
+    (dispatch, getState) => {
+        if (
+            removedDevice.serialNumber ===
+            getState().device.selectedSerialNumber
+        ) {
+            onDeviceDeselected();
+        }
+
+        if (
+            removedDevice.serialNumber ===
+                getState().device.selectedSerialNumber &&
+            !getState().deviceAutoSelect.waitForDevice
+        ) {
+            dispatch(closeDeviceSetupDialog());
+        }
+
+        if (!getState().deviceAutoSelect.arrivedButWrongWhen) {
+            dispatch(removeDevice(removedDevice));
+            onDeviceDisconnected(removedDevice);
+        }
+    };
 /*
  * Starts watching for devices with the given traits. See the nrf-device-lib
  * library for available traits. Whenever devices are attached/detached, this
@@ -148,31 +175,9 @@ export const startWatchingDevices =
         onDeviceDisconnected: (device: Device) => void,
         onDeviceDeselected: () => void,
         doSelectDevice: (device: Device, autoReselected: boolean) => void
-    ) =>
-    async (dispatch: AppDispatch, getState: () => RootState) => {
+    ): AppThunk<RootState, Promise<void>> =>
+    async (dispatch, getState) => {
         const updateDeviceList = (event: HotplugEvent) => {
-            const removeDeviceFromList = (remove: Device) => {
-                if (
-                    remove.serialNumber ===
-                    getState().device.selectedSerialNumber
-                ) {
-                    onDeviceDeselected();
-                }
-
-                if (
-                    remove.serialNumber ===
-                        getState().device.selectedSerialNumber &&
-                    !getState().deviceAutoSelect.waitForDevice
-                ) {
-                    dispatch(closeDeviceSetupDialog());
-                }
-
-                if (!getState().deviceAutoSelect.arrivedButWrongWhen) {
-                    dispatch(removeDevice(remove));
-                    onDeviceDisconnected(remove);
-                }
-            };
-
             switch (event.event_type) {
                 case 'NRFDL_DEVICE_EVENT_ARRIVED':
                     if (!event.device) {
@@ -277,11 +282,11 @@ export const startWatchingDevices =
                                         'Wait For Device was successfully'
                                     );
 
-                                    if (waitForDevice.once) {
-                                        dispatch(clearWaitForDevice());
-                                    } else {
-                                        dispatch(clearWaitForDeviceTimeout());
-                                    }
+                                    dispatch(
+                                        clearWaitForDeviceTimeout(
+                                            waitForDevice.once
+                                        )
+                                    );
 
                                     if (waitForDevice.onSuccess)
                                         waitForDevice.onSuccess(
@@ -315,20 +320,35 @@ export const startWatchingDevices =
                                         dispatch(
                                             initAutoReconnectTimeout(
                                                 () =>
-                                                    removeDeviceFromList(
-                                                        device
+                                                    dispatch(
+                                                        removeDeviceFromList(
+                                                            device,
+                                                            onDeviceDeselected,
+                                                            onDeviceDisconnected
+                                                        )
                                                     ),
-                                                getState().deviceAutoSelect
-                                                    .waitForDevice
+                                                waitForDevice
                                             )
                                         );
                                     } else {
-                                        removeDeviceFromList(device);
+                                        dispatch(
+                                            removeDeviceFromList(
+                                                device,
+                                                onDeviceDeselected,
+                                                onDeviceDisconnected
+                                            )
+                                        );
                                     }
 
                                     dispatch(setDisconnectedTime(Date.now()));
                                 } else {
-                                    removeDeviceFromList(device);
+                                    dispatch(
+                                        removeDeviceFromList(
+                                            device,
+                                            onDeviceDeselected,
+                                            onDeviceDisconnected
+                                        )
+                                    );
                                 }
                             }
                         });
@@ -377,6 +397,13 @@ const getAutoSelectDeviceCLISerial = () => {
     const { argv } = process;
     const serialIndex = argv.findIndex(arg => arg === '--deviceSerial');
     return serialIndex > -1 ? argv[serialIndex + 1] : undefined;
+};
+
+export const clearWaitForDevice = (): AppThunk => (dispatch, getState) => {
+    if (getState().deviceAutoSelect.autoReconnectTimeout) {
+        getState().deviceAutoSelect.onCancelTimeout?.();
+    }
+    dispatch(clearWaitForDeviceTimeout(true));
 };
 
 /**