+++ src/Device/sdfuOperations.ts
@@ -11,7 +11,7 @@ import fs from 'fs';
 import MemoryMap from 'nrf-intel-hex';
 
 import logger from '../logging';
-import { AppDispatch } from '../store';
+import { AppThunk, RootState } from '../store';
 import { getAppFile } from '../utils/appDirs';
 import { setWaitForDevice } from './deviceAutoSelectSlice';
 import { getDeviceLibContext } from './deviceLibWrapper';
@@ -87,8 +87,8 @@ const updateBootloader =
         onSuccess: (device: Device) => void,
         onFail: (reason?: unknown) => void,
         onProgress: (progress: number, message?: string) => void
-    ) =>
-    async (dispatch: AppDispatch) => {
+    ): AppThunk =>
+    async dispatch => {
         logger.info(`Update Bootloader ${device}`);
         const zip = new AdmZip(getAppFile(LATEST_BOOTLOADER));
         const zipBuffer = zip.toBuffer();
@@ -153,8 +153,8 @@ const switchToDeviceMode =
         mcuState: nrfDeviceLib.MCUState,
         onSuccess: (device: Device) => void,
         onFail: (reason?: unknown) => void
-    ) =>
-    (dispatch: AppDispatch) => {
+    ): AppThunk =>
+    dispatch => {
         nrfDeviceLib
             .deviceControlSetMcuState(
                 getDeviceLibContext(),
@@ -183,8 +183,8 @@ export const switchToBootloaderMode =
         device: Device,
         onSuccess: (device: Device) => void,
         onFail: (reason?: unknown) => void
-    ) =>
-    (dispatch: AppDispatch) => {
+    ): AppThunk =>
+    dispatch => {
         if (!isDeviceInDFUBootloader(device)) {
             dispatch(
                 switchToDeviceMode(
@@ -210,8 +210,8 @@ export const switchToApplicationMode =
         device: Device,
         onSuccess: (device: Device) => void,
         onFail: (reason?: unknown) => void
-    ) =>
-    (dispatch: AppDispatch) => {
+    ): AppThunk =>
+    dispatch => {
         dispatch(
             switchToDeviceMode(
                 device,
@@ -253,8 +253,8 @@ const askAndUpdateBootloader =
         onSuccess: (device: Device) => void,
         onFail: (reason?: unknown) => void,
         onProgress: (progress: number, message?: string) => void
-    ) =>
-    (dispatch: AppDispatch) => {
+    ): AppThunk =>
+    dispatch => {
         dispatch(
             switchToBootloaderMode(
                 device,
@@ -413,8 +413,8 @@ const programInDFUBootloader =
         onProgress: (progress: number, message?: string) => void,
         onSuccess: (device: Device) => void,
         onFail: (reason?: unknown) => void
-    ) =>
-    async (dispatch: AppDispatch) => {
+    ): AppThunk<RootState, Promise<void>> =>
+    async dispatch => {
         logger.debug(
             `${device.serialNumber} on ${device.serialport?.comName} is now in DFU-Bootloader...`
         );
@@ -526,8 +526,8 @@ const programDeviceWithFw =
         device: Device,
         selectedFw: DfuEntry,
         onProgress: (progress: number, message?: string) => void
-    ) =>
-    (dispatch: AppDispatch) =>
+    ): AppThunk<RootState, Promise<Device>> =>
+    dispatch =>
         new Promise<Device>((resolve, reject) => {
             const action = (d: Device) => {
                 dispatch(
@@ -560,7 +560,7 @@ export const sdfuDeviceSetup = (
         dfuFirmware.map(firmwareOption => ({
             key: firmwareOption.key,
             description: firmwareOption.description,
-            programDevice: onProgress => (dispatch: AppDispatch) =>
+            programDevice: onProgress => dispatch =>
                 dispatch(
                     programDeviceWithFw(device, firmwareOption, onProgress)
                 ),
@@ -590,7 +590,7 @@ export const sdfuDeviceSetup = (
                 });
             }
         }),
-    tryToSwitchToApplicationMode: device => (dispatch: AppDispatch) =>
+    tryToSwitchToApplicationMode: device => dispatch =>
         new Promise<Device>((resolve, reject) => {
             dispatch(switchToApplicationMode(device, resolve, reject));
         }),